refactor!: `GeneralConfigProvider`
Removes a few hook and contexts from being exposed at the top level.
