+++ README.md
@@ -125,8 +125,6 @@ The transaction categorization component handles displaying both categorized tra
 <BankTransactions asWidget />
 ```
 
-Optional properties and `useBankTransactionsContext` give more control over the transactions list:
-
 ```tsx
 /** Using props */
 <BankTransactions
@@ -134,13 +132,6 @@ Optional properties and `useBankTransactionsContext` give more control over the 
     amount: { min: 0, max: 100 },
   }}
 />
-
-/** Using hook */
-const { setFilters } = useBankTransactionsContext()
-
-setFilters({ amount: { min: 0, max: 10000 } })
-
-<BankTransactions />
 ```
 
 ### Reporting
