--- src/contexts/LayerContext/index.tsx
@@ -1,1 +1,0 @@
-export { LayerContext, useLayerContext } from './LayerContext'
