+++ src/hooks/useChartOfAccounts/useChartOfAccounts.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react'
 import { Layer } from '../../api/layer'
 import { NORMALITY_OPTIONS } from '../../components/ChartOfAccountsForm/constants'
-import { useLayerContext } from '../../contexts/LayerContext'
+import { useLayerContext } from '../../contexts/LayerContext/LayerContext'
 import { FormError, DateRange, Direction, NewAccount } from '../../types'
 import {
   EditAccount,
@@ -11,8 +11,8 @@ import { BaseSelectOption, DataModel } from '../../types/general'
 import { endOfMonth, formatISO, startOfMonth } from 'date-fns'
 import useSWR from 'swr'
 import { useAuth } from '../useAuth'
-import { useEnvironment } from '../../providers/Environment/EnvironmentInputProvider'
-import { useBusinessId } from '../../providers/BusinessProvider/BusinessInputProvider'
+import { useEnvironment } from '../../providers/environment/EnvironmentInputProvider'
+import { useBusinessId } from '../../providers/business/BusinessInputProvider'
 
 const validate = (formData?: ChartOfAccountsForm) => {
   const errors: FormError[] = []
